+++ app/controllers/api/v1/characters_controller.rb
@@ -3,7 +3,7 @@
 module Api
   module V1
     class CharactersController < BaseController
-      before_action :authenticate_player
+      before_action :authenticate_player, except: :show
       before_action :set_character, only: %i[show update destroy]
 
       def show