+++ app/javascript/lca/components/characters/editors/basicsEditor.jsx
@@ -1,15 +1,18 @@
 import React from 'react'
 import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
 
+import Checkbox from 'material-ui/Checkbox'
+import { FormControlLabel } from 'material-ui/Form'
 import TextField from 'material-ui/TextField'
 
 import BlockPaper from '../../generic/blockPaper.jsx'
 import RatingField from '../../generic/RatingField.jsx'
-
+import { canIDeleteCharacter } from '../../../selectors'
 import { ESSENCE_MIN, ESSENCE_MAX } from '../../../utils/constants.js'
 
 function BasicsEditor(props) {
-  const { character, onChange, onBlur, onRatingChange } = props
+  const { character, onChange, onBlur, onRatingChange, onCheck, showPublicCheckbox } = props
 
   return <BlockPaper>
     <TextField name="name" value={ character.name }
@@ -26,6 +29,17 @@ function BasicsEditor(props) {
       label="Description" margin="dense"
       multiline fullWidth rows={ 2 } rowsMax={ 10 }
       onChange={ onChange } onBlur={ onBlur } />
+
+    { showPublicCheckbox &&
+      <FormControlLabel
+        label="Publicly Viewable"
+        control={
+          <Checkbox name="public" checked={ character.public }
+            onChange={ onCheck }
+          />
+        }
+      />
+    }
   </BlockPaper>
 
 }
@@ -34,6 +48,11 @@ BasicsEditor.propTypes = {
   onChange: PropTypes.func.isRequired,
   onBlur: PropTypes.func.isRequired,
   onRatingChange: PropTypes.func.isRequired,
+  onCheck: PropTypes.func.isRequired,
+  showPublicCheckbox: PropTypes.bool,
 }
+const mapStateToProps = (state, props) => ({
+  showPublicCheckbox: canIDeleteCharacter(state, props.character.id),
+})
 
-export default BasicsEditor
+export default connect(mapStateToProps)(BasicsEditor)