+++ app/javascript/lca/selectors/battlegroup.js
@@ -15,6 +15,9 @@ export const getAttacksForBattlegroup = createCachedSelector(
 export const canIEditBattlegroup = createSelector(
   [getCurrentPlayer, getSpecificBattlegroup, getState],
   (player, character, state) => {
+    if (character === undefined)
+      return false
+
     if (player.id === character.player_id)
       return true