+++ db/migrate/20180406052142_allow_publicly_viewable_characters.rb
@@ -1,0 +1,7 @@
+class AllowPubliclyViewableCharacters < ActiveRecord::Migration[5.1]
+  def change
+    add_column :characters,   :public, :boolean, default: false
+    add_column :qcs,          :public, :boolean, default: false
+    add_column :battlegroups, :public, :boolean, default: false
+  end
+end