+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180401235315) do
+ActiveRecord::Schema.define(version: 20180406052142) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -50,6 +50,7 @@ ActiveRecord::Schema.define(version: 20180401235315) do
     t.integer "chronicle_sort_order", default: 0
     t.boolean "in_combat", default: false
     t.boolean "has_acted", default: false
+    t.boolean "public", default: false
     t.index ["chronicle_id"], name: "index_battlegroups_on_chronicle_id"
     t.index ["player_id"], name: "index_battlegroups_on_player_id"
   end
@@ -163,6 +164,7 @@ ActiveRecord::Schema.define(version: 20180401235315) do
     t.text "notes", default: ""
     t.boolean "houserules", default: false
     t.json "bp_log", default: []
+    t.boolean "public", default: false
     t.index ["chronicle_id"], name: "index_characters_on_chronicle_id"
     t.index ["player_id"], name: "index_characters_on_player_id"
   end
@@ -369,6 +371,7 @@ ActiveRecord::Schema.define(version: 20180401235315) do
     t.boolean "in_combat", default: false
     t.boolean "has_acted", default: false
     t.text "notes", default: ""
+    t.boolean "public", default: false
     t.index ["chronicle_id"], name: "index_qcs_on_chronicle_id"
     t.index ["player_id"], name: "index_qcs_on_player_id"
   end